add req.resume()

+++ logic/route/upload.js
@@ -30,20 +30,22 @@ module.exports = (config, app, common, route) => {
       app.console.debug(`[${user.username}] Started upload`)
       req.on('close', async () => {
         if (!uploadState.complete) {
-          app.console.debug(`[${user.username}] Upload aborted removing ${uploadState.files.paths.length} files`, 'red')
-          await common.asyncForEach(uploadState.files.paths, async partial => {
-            if (partial.stream !== null) {
-              uploadState.abort = true
-              partial.stream.end()
-            }
-            return new Promise((resolve, reject) => {
-              fs.unlink(partial.file, () => {
-                uploadState.files.removed += 1
-                return resolve()
+          if (uploadState.files.paths.length > 0) {
+            app.console.debug(`[${user.username}] Upload aborted removing ${uploadState.files.paths.length} files`, 'red')
+            await common.asyncForEach(uploadState.files.paths, async partial => {
+              if (partial.stream !== null) {
+                uploadState.abort = true
+                partial.stream.end()
+              }
+              return new Promise((resolve, reject) => {
+                fs.unlink(partial.file, () => {
+                  uploadState.files.removed += 1
+                  return resolve()
+                })
               })
             })
-          })
-          app.console.debug(`${uploadState.files.removed} Files removed`, 'red')
+            app.console.debug(`${uploadState.files.removed} Files removed`, 'red')
+          }
         } else {
           app.console.debug(`[${user.username}] Upload Completed`)
         }
@@ -166,6 +168,7 @@ module.exports = (config, app, common, route) => {
             if (!signature.mimetype.includes(fileinfo.type)) {
               app.console.debug(`Upload of '${fileinfo.id}' rejected file magic is invaild ('${signature.mimetype}')`, 'red')
               req.unpipe(req.busboy)
+              req.resume()
               return res.status(415).json({
                 file: filename,
                 status: 415,
@@ -190,7 +193,8 @@ module.exports = (config, app, common, route) => {
               Bug with unpiping causing no response to be returned by the server
               https://github.com/mscdex/busboy/issues/209
             */
-           // req.unpipe(req.busboy)
+            req.unpipe(req.busboy)
+            req.resume()
             return res.status(413).json({
               file: 'filename',
               status: 413,
@@ -201,6 +205,7 @@ module.exports = (config, app, common, route) => {
         } else {
           app.console.debug(`Upload of '${fileinfo.id}' aborted invaild filetype`, 'red')
           req.unpipe(req.busboy)
+          req.resume()
           return res.status(415).json({
             file: filename,
             status: 415,
